+++ include/openssl/sslerr_legacy.h
@@ -461,6 +461,26 @@ OSSL_DEPRECATEDIN_3_0 int ERR_load_SSL_strings(void);
 #  define SSL_F_WRITE_STATE_MACHINE                        0
 # endif
 
+# ifndef OPENSSL_NO_DEPRECATED_4_0
+
+#  define SSL_R_SSL3_EXT_INVALID_MAX_FRAGMENT_LENGTH       232
+#  define SSL_R_SSL3_EXT_INVALID_SERVERNAME                319
+#  define SSL_R_SSL3_EXT_INVALID_SERVERNAME_TYPE           320
+#  define SSL_R_SSL3_SESSION_ID_TOO_LONG                   300
+#  define SSL_R_SSLV3_ALERT_BAD_CERTIFICATE                1042
+#  define SSL_R_SSLV3_ALERT_BAD_RECORD_MAC                 1020
+#  define SSL_R_SSLV3_ALERT_CERTIFICATE_EXPIRED            1045
+#  define SSL_R_SSLV3_ALERT_CERTIFICATE_REVOKED            1044
+#  define SSL_R_SSLV3_ALERT_CERTIFICATE_UNKNOWN            1046
+#  define SSL_R_SSLV3_ALERT_DECOMPRESSION_FAILURE          1030
+#  define SSL_R_SSLV3_ALERT_HANDSHAKE_FAILURE              1040
+#  define SSL_R_SSLV3_ALERT_ILLEGAL_PARAMETER              1047
+#  define SSL_R_SSLV3_ALERT_NO_CERTIFICATE                 1041
+#  define SSL_R_SSLV3_ALERT_UNEXPECTED_MESSAGE             1010
+#  define SSL_R_SSLV3_ALERT_UNSUPPORTED_CERTIFICATE        1043
+
+# endif
+
 # ifdef  __cplusplus
 }
 # endif